SPEECH_TO_TEXT/requirement.txt
python-dotenv
deepgram-sdk==3.*
pyaudio